+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [v1.2.0]
+
+- Runs Liquidhaskell on save automatically (if enabled)
+- Added configuration to add custom command used to run LH
+
 ## [v1.0.0]
 
 - Initial release - shows errors in vscode's diagnostics, works only for workspaces (folders) and not single files.